-m KSKSMGMT-78
Former-commit-id: b707e6dae7b79d3dd1654ccb0b661362498f75bc

+++ management/server/broker/manager-persistence/src/main/java/org/safehaus/kiskis/mgmt/server/persistence/Persistence.java
@@ -3,17 +3,13 @@ package org.safehaus.kiskis.mgmt.server.persistence;
 import com.datastax.driver.core.*;
 import org.safehaus.kiskis.mgmt.shared.protocol.*;
 import org.safehaus.kiskis.mgmt.shared.protocol.api.PersistenceInterface;
+import org.safehaus.kiskis.mgmt.shared.protocol.enums.RequestType;
 import org.safehaus.kiskis.mgmt.shared.protocol.enums.ResponseType;
 import org.safehaus.kiskis.mgmt.shared.protocol.enums.TaskStatus;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import org.safehaus.kiskis.mgmt.shared.protocol.enums.RequestType;
 
 /**
  * Created with IntelliJ IDEA. User: daralbaev Date: 11/7/13 Time: 10:57 PM
@@ -23,7 +19,6 @@ public class Persistence implements PersistenceInterface {
     private static final Logger LOG = Logger.getLogger(Persistence.class.getName());
     private Cluster cluster;
     private Session session;
-    private long requestsequencenumber = 0l;
     private String cassandraHost;
     private String cassandraKeyspace;
     private int cassandraPort;
@@ -52,7 +47,7 @@ public class Persistence implements PersistenceInterface {
                     new Date(System.currentTimeMillis() - to * 60 * 1000)));
             for (Row row : rs) {
                 Agent agent = new Agent();
-                agent.setUuid(row.getString("uuid"));
+                agent.setUuid(row.getUUID("uuid"));
                 agent.setHostname(row.getString("hostname"));
                 agent.setIsLXC(row.getBool("islxc"));
                 agent.setLastHeartbeat(row.getDate("lastheartbeat"));
@@ -77,7 +72,7 @@ public class Persistence implements PersistenceInterface {
                     new Date(System.currentTimeMillis() - freshness * 60 * 1000)));
             for (Row row : rs) {
                 Agent agent = new Agent();
-                agent.setUuid(row.getString("uuid"));
+                agent.setUuid(row.getUUID("uuid"));
                 agent.setHostname(row.getString("hostname"));
                 agent.setIsLXC(row.getBool("islxc"));
                 agent.setLastHeartbeat(row.getDate("lastheartbeat"));
@@ -124,7 +119,7 @@ public class Persistence implements PersistenceInterface {
     @Override
     public boolean saveCommand(Command command) {
         try {
-            String cql = "insert into request (source, requestsequencenumber, type, uuid, taskuuid, "
+            String cql = "insert into requests (source, reqseqnum, type, agentuuid, taskuuid, "
                     + "workingdirectory, program, outputredirectionstdout, outputredirectionstderr, "
                     + "stdoutpath, erroutpath, runsas, args, environment, pid, timeout) "
                     + "values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);";
@@ -132,7 +127,7 @@ public class Persistence implements PersistenceInterface {
             BoundStatement boundStatement = new BoundStatement(stmt);
 
             Request request = command.getCommand();
-            ResultSet rs = session.execute(boundStatement.bind(request.getSource(), requestsequencenumber++,
+            ResultSet rs = session.execute(boundStatement.bind(request.getSource(), request.getRequestSequenceNumber(),
                     request.getType() + "", request.getUuid(), request.getTaskUuid(), request.getWorkingDirectory(),
                     request.getProgram(), request.getStdOut() + "", request.getStdErr() + "",
                     request.getStdOutPath(), request.getStdErrPath(), request.getRunAs(), request.getArgs(),
@@ -154,15 +149,17 @@ public class Persistence implements PersistenceInterface {
     public boolean saveResponse(Response response) {
         try {
 
-            String cql = "insert into response (uuid, taskuuid, responsesequencenumber, exitcode, errout, hostname, ips, "
-                    + "macaddress, pid, requestsequencenumber, responsetype, source, stdout, islxc) "
+            String cql = "insert into responses (agentuuid, taskuuid, resseqnum, exitcode, errout, hostname, ips, "
+                    + "macaddress, pid, reqseqnum, responsetype, source, stdout, islxc) "
                     + "values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);";
             PreparedStatement stmt = session.prepare(cql);
             BoundStatement boundStatement = new BoundStatement(stmt);
             ResultSet rs = session.execute(boundStatement.bind(response.getUuid(), response.getTaskUuid(),
-                    (Long) response.getResponseSequenceNumber() == null ? -1l : response.getResponseSequenceNumber(),
+                    response.getResponseSequenceNumber() == null ? -1 : response.getResponseSequenceNumber(),
                     response.getExitCode(), response.getStdErr(), response.getHostname(), response.getIps(),
-                    response.getMacAddress(), response.getPid(), response.getRequestSequenceNumber(), response.getType().toString(),
+                    response.getMacAddress(), response.getPid(),
+                    response.getRequestSequenceNumber() == null ? -1 : response.getRequestSequenceNumber(),
+                    response.getType().toString(),
                     response.getSource(), response.getStdOut(), response.isLxc));
             return true;
         } catch (Exception ex) {
@@ -172,32 +169,32 @@ public class Persistence implements PersistenceInterface {
     }
 
     @Override
-    public List<Response> getResponses(String taskuuid, Long requestSequenceNumber) {
+    public List<Response> getResponses(UUID taskuuid, Integer requestSequenceNumber) {
         List<Response> list = new ArrayList<Response>();
         try {
-            String cql = "select * from response " +
+            String cql = "select * from responses " +
                     "WHERE taskuuid = ? " +
-                    "and requestsequencenumber = ? and responsesequencenumber >= 0 " +
-                    "ORDER BY requestsequencenumber, responsesequencenumber";
+                    "and reqseqnum = ? and resseqnum >= 0 " +
+                    "ORDER BY reqseqnum, resseqnum";
             PreparedStatement stmt = session.prepare(cql);
             BoundStatement boundStatement = new BoundStatement(stmt);
             ResultSet rs = session.execute(boundStatement.bind(taskuuid, requestSequenceNumber));
             for (Row row : rs) {
                 Response response = new Response();
                 response.setType(ResponseType.valueOf(row.getString("responsetype")));
-                response.setUuid(row.getString("uuid"));
+                response.setUuid(row.getUUID("agentuuid"));
                 response.setExitCode(row.getInt("exitcode"));
                 response.setStdOut(row.getString("stdout"));
                 response.setHostname(row.getString("hostname"));
                 response.setIps(row.getList("ips", String.class));
                 response.setMacAddress(row.getString("macaddress"));
                 response.setPid(row.getInt("pid"));
-                response.setRequestSequenceNumber(row.getLong("requestsequencenumber"));
-                response.setResponseSequenceNumber(row.getLong("responsesequencenumber"));
+                response.setRequestSequenceNumber(row.getInt("reqseqnum"));
+                response.setResponseSequenceNumber(row.getInt("resseqnum"));
                 response.setSource(row.getString("source"));
                 response.setStdErr(row.getString("errout"));
                 response.setStdOut(row.getString("stdout"));
-                response.setTaskUuid(row.getString("taskuuid"));
+                response.setTaskUuid(row.getUUID("taskuuid"));
                 response.setType(ResponseType.valueOf(row.getString("responsetype")));
                 list.add(response);
             }
@@ -241,10 +238,12 @@ public class Persistence implements PersistenceInterface {
     @Override
     public boolean removeAgent(Agent agent) {
         try {
-            String cql = "delete from agents where uuid = ?";
-            PreparedStatement stmt = session.prepare(cql);
-            BoundStatement boundStatement = new BoundStatement(stmt);
-            ResultSet rs = session.execute(boundStatement.bind(agent.getUuid()));
+            if (agent != null) {
+                String cql = "delete from agents where uuid = ?";
+                PreparedStatement stmt = session.prepare(cql);
+                BoundStatement boundStatement = new BoundStatement(stmt);
+                ResultSet rs = session.execute(boundStatement.bind(agent.getUuid()));
+            }
             return true;
 
         } catch (Exception ex) {
@@ -264,15 +263,15 @@ public class Persistence implements PersistenceInterface {
     @Override
     public String saveTask(Task task) {
         try {
-            String cql = "insert into tasks (uid, description, status) "
+            String cql = "insert into tasks (uuid, description, status) "
                     + "values (?, ?, ?);";
             PreparedStatement stmt = session.prepare(cql);
 
             BoundStatement boundStatement = new BoundStatement(stmt);
 
-            session.execute(boundStatement.bind(task.getUid(), task.getDescription(), task.getTaskStatus().toString()));
+            session.execute(boundStatement.bind(task.getUuid(), task.getDescription(), task.getTaskStatus().toString()));
 
-            return task.getUid().toString();
+            return task.getUuid().toString();
 
         } catch (Exception ex) {
             LOG.log(Level.SEVERE, "Error in saveTask", ex);
@@ -281,10 +280,10 @@ public class Persistence implements PersistenceInterface {
     }
 
     @Override
-    public List<Request> getRequests(String taskuuid) {
+    public List<Request> getRequests(UUID taskuuid) {
         List<Request> list = new ArrayList<Request>();
         try {
-            ResultSet rs = session.execute("select * from request");
+            ResultSet rs = session.execute("select * from requests");
             for (Row row : rs) {
                 Request request = new Request();
                 request.setProgram(row.getString("program"));
@@ -292,17 +291,17 @@ public class Persistence implements PersistenceInterface {
                 request.setEnvironment(row.getMap("environment", String.class, String.class));
                 request.setPid(row.getInt("pid"));
                 request.setProgram(row.getString("program"));
-                request.setRequestSequenceNumber(row.getLong("requestsequencenumber"));
+                request.setRequestSequenceNumber(row.getInt("reqseqnum"));
                 request.setRunAs(row.getString("runsas"));
                 request.setSource(row.getString("source"));
                 request.setStdErr(OutputRedirection.valueOf(row.getString("outputredirectionstderr")));
                 request.setStdErrPath(row.getString("erroutpath"));
                 request.setStdOut(OutputRedirection.valueOf(row.getString("outputredirectionstdout")));
                 request.setStdOutPath(row.getString("stdoutpath"));
-                request.setTaskUuid(row.getString("taskuuid"));
-                request.setTimeout(row.getLong("timeout"));
+                request.setTaskUuid(row.getUUID("taskuuid"));
+                request.setTimeout(row.getInt("timeout"));
                 request.setType(RequestType.valueOf(row.getString("type")));
-                request.setUuid(row.getString("uuid"));
+                request.setUuid(row.getUUID("agentuuid"));
                 request.setWorkingDirectory(row.getString("workingdirectory"));
                 list.add(request);
             }
@@ -320,7 +319,7 @@ public class Persistence implements PersistenceInterface {
             ResultSet rs = session.execute("select * from tasks");
             for (Row row : rs) {
                 Task task = new Task();
-                task.setUid(row.getUUID("uid"));
+                task.setUuid(row.getUUID("uuid"));
                 task.setDescription(row.getString("description"));
                 task.setTaskStatus(TaskStatus.valueOf(row.getString("status")));
                 list.add(task);
@@ -335,8 +334,8 @@ public class Persistence implements PersistenceInterface {
     public boolean truncateTables() {
         try {
             session.execute("truncate tasks");
-            session.execute("truncate request");
-            session.execute("truncate response");
+            session.execute("truncate requests");
+            session.execute("truncate responses");
         } catch (Exception ex) {
             LOG.log(Level.SEVERE, "Error in getTasks", ex);
             return false;
